lint: enable detekt rule: LargeClass
