+++ backend/src/test/kotlin/com/moneat/services/NetworkDevicesTest.kt
@@ -14,8 +14,6 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-@file:Suppress("LargeClass")
-
 package com.moneat.services
 
 import com.moneat.config.ClickHouseClient
